+++ lttng/org.eclipse.linuxtools.tmf.core/src/org/eclipse/linuxtools/tmf/core/ctfadaptor/CtfTmfEventField.java
@@ -215,7 +215,7 @@ final class CTFIntegerField extends CtfTmfEventField {
      *            Is the value signed or not
      */
     CTFIntegerField(String name, long longValue, int base, boolean signed) {
-        super(name, Long.valueOf(longValue), null);
+        super(name, longValue, null);
         this.signed = signed;
         this.base = base;
     }
@@ -227,7 +227,7 @@ final class CTFIntegerField extends CtfTmfEventField {
 
     @Override
     public String getFormattedValue() {
-        return IntegerDefinition.formatNumber(getValue().longValue(), base, signed);
+        return IntegerDefinition.formatNumber(getValue(), base, signed);
     }
 
 }
@@ -359,7 +359,7 @@ final class CTFFloatField extends CtfTmfEventField {
      *            The name of this field
      */
     protected CTFFloatField(String name, double value) {
-        super(name, Double.valueOf(value), null);
+        super(name, value, null);
     }
 
     @Override
@@ -386,7 +386,7 @@ final class CTFEnumField extends CtfTmfEventField {
      */
     CTFEnumField(String name, CtfEnumPair enumValue) {
         super(name, new CtfEnumPair(enumValue.getFirst(),
-                enumValue.getSecond()), null);
+                enumValue.getSecond().longValue()), null);
     }
 
     @Override
