+++ lttng/org.eclipse.linuxtools.tmf.ui.tests/widgetStubs/org/eclipse/linuxtools/tmf/ui/tests/uml2sd/trace/TmfUml2SDTestTrace.java
@@ -69,7 +69,7 @@ public class TmfUml2SDTestTrace implements ITmfEventParser {
 
         long location = 0;
         if (context != null) {
-            location = ((Long) context.getLocation().getLocationInfo()).longValue();
+            location = (Long) context.getLocation().getLocationInfo();
         }
 
         try {
